job-tracker-simple/python/requirements.txt
# Job Tracker Simple - Python Requirements
# Version: 1.0

# === SUPABASE CLIENT ===
supabase==2.4.0
postgrest==0.16.8

# === UTILITAIRES ===
python-dotenv==1.0.1
requests==2.32.3

# === LinkedIn API (réutilisation du workflow existant) ===
# Ces dépendances sont déjà installées dans le projet principal linkedin-mcp
# linkedin-api==2.3.1
# beautifulsoup4==4.13.4
# lxml==5.4.0

# === WEB SCRAPING ===
playwright==1.48.0

# === DÉVELOPPEMENT ET DEBUG ===
python-dateutil==2.9.0